refactor: move main source, decouple lib
* moved main anime-dl source to `/src`
* keep `check-executable` lib util separate
