--- lib/constants/urls.ts
@@ -1,3 +1,0 @@
-// Gogo Urls
-export const GOGO_ROOT = "https://gogoanime.cm";
-export const GOGO_ENCRYPT_AJAX = "https://gogoplay.io/encrypt-ajax.php";
