+++ lib/index.ts
@@ -1,0 +1,5 @@
+import checkExecutableSync from "./check-executable";
+
+export default {
+  checkExecutableSync,
+};
