+++ src/constants/urls.ts
@@ -1,0 +1,3 @@
+// Gogo Urls
+export const GOGO_ROOT = "https://gogoanime.gg/";
+export const GOGO_ENCRYPT_AJAX = "https://gogoplay.io/encrypt-ajax.php/";
