+++ src/locales.ts
